+++ modules/tinymce/src/core/main/ts/caret/FakeCaret.ts
@@ -9,7 +9,6 @@ import { Singleton } from '@ephox/katamari';
 import { PlatformDetection } from '@ephox/sand';
 import { SelectorFilter, SugarElement } from '@ephox/sugar';
 
-import DomQuery from '../api/dom/DomQuery';
 import Editor from '../api/Editor';
 import * as Settings from '../api/Settings';
 import Delay from '../api/util/Delay';
@@ -111,6 +110,7 @@ export const FakeCaret = (editor: Editor, root: HTMLElement, isBlock: (node: Nod
   let caretContainerNode: Node | null;
   const rootBlock = Settings.getForcedRootBlock(editor);
   const caretBlock = rootBlock.length > 0 ? rootBlock : 'p';
+  const dom = editor.dom;
 
   const show = (before: boolean, element: Element): Range | null => {
     let rng: Range;
@@ -124,13 +124,15 @@ export const FakeCaret = (editor: Editor, root: HTMLElement, isBlock: (node: Nod
     if (isBlock(element)) {
       caretContainerNode = CaretContainer.insertBlock(caretBlock, element, before);
       const clientRect = getAbsoluteClientRect(root, element, before);
-      DomQuery(caretContainerNode).css('top', clientRect.top);
+      dom.setStyle(caretContainerNode, 'top', clientRect.top);
 
-      const caret = DomQuery<HTMLElement>('<div class="mce-visual-caret" data-mce-bogus="all"></div>').css({ ...clientRect }).appendTo(root)[0];
+      const caret = dom.create('div', { 'class': 'mce-visual-caret', 'data-mce-bogus': 'all' });
+      dom.setStyles(caret, { ...clientRect });
+      dom.add(root, caret);
       lastVisualCaret.set({ caret, element, before });
 
       if (before) {
-        DomQuery(caret).addClass('mce-visual-caret-before');
+        dom.addClass(caret, 'mce-visual-caret-before');
       }
       startBlink();
 
@@ -166,7 +168,7 @@ export const FakeCaret = (editor: Editor, root: HTMLElement, isBlock: (node: Nod
     }
 
     lastVisualCaret.on((caretState) => {
-      DomQuery(caretState.caret).remove();
+      dom.remove(caretState.caret);
       lastVisualCaret.clear();
     });
 
@@ -178,18 +180,20 @@ export const FakeCaret = (editor: Editor, root: HTMLElement, isBlock: (node: Nod
 
   const startBlink = () => {
     cursorInterval = Delay.setInterval(() => {
-      if (hasFocus()) {
-        DomQuery('div.mce-visual-caret', root).toggleClass('mce-visual-caret-hidden');
-      } else {
-        DomQuery('div.mce-visual-caret', root).addClass('mce-visual-caret-hidden');
-      }
+      lastVisualCaret.on((caretState) => {
+        if (hasFocus()) {
+          dom.toggleClass(caretState.caret, 'mce-visual-caret-hidden');
+        } else {
+          dom.addClass(caretState.caret, 'mce-visual-caret-hidden');
+        }
+      });
     }, 500);
   };
 
   const reposition = () => {
     lastVisualCaret.on((caretState) => {
       const clientRect = getAbsoluteClientRect(root, caretState.element, caretState.before);
-      DomQuery(caretState.caret).css({ ...clientRect });
+      dom.setStyles(caretState.caret, { ...clientRect });
     });
   };